// src/DVLD.Client/src/app/features/people/models/edit-person.ts
export interface EditPersonName {
  firstName: string;
  secondName: string;
  thirdName: string;
  lastName: string;
  motherName: string;
}

export interface EditPersonContactInfo {
  phone: string;
  altPhone: string;
}

export interface EditPersonalInfo {
  dateOfBirth: string;
  nationalityCountryCode: string;
}